CSV functionality
added csv functionality. Had to add some additional code to get rid of the quotation marks when reading the line.

Took out JSON capabilities since i need to get started on the github action side of things. This can be improved (see the FIXMEs in the code) but this is pretty solid AND A W IS A W